+++ xapian-core/spelling/spelling_splitter.h
@@ -92,7 +92,7 @@ class SpellingSplitter : public SpellingBase {
     bool request_word_exists(const word_splitter_data& data, unsigned word_start,
                              unsigned word_end, string& word) const;
 
-    unsigned get_sort_distance(const word_splitter_temp& temp,
+    double get_sort_distance(const word_splitter_temp& temp,
                                word_splitter_value first,
                                word_splitter_value second) const;
